Move custom app token factory to service layer
