+++ tests/conftest.py
@@ -16,10 +16,11 @@ from sqlalchemy.orm import sessionmaker
 from sqlalchemy_utils import create_database, database_exists, drop_database
 
 from app.clients.db.session import get_db
-from app.core import custom_app, security
+from app.core import security
 from app.core.aws import S3Client, get_s3_client
-from app.core.custom_app import AppTokenFactory
 from app.main import app
+from app.service import custom_app
+from app.service.custom_app import AppTokenFactory
 
 
 @pytest.fixture(scope="function")
